requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
attrs==21.2.0
Automat==20.2.0
autopep8==2.0.4
Babel==2.8.0
bcrypt==3.2.0
black==23.11.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==23.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
greenlet==1.1.2
h11==0.14.0
httplib2==0.20.2
httptools==0.6.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.0
MarkupSafe==2.0.1
more-itertools==8.10.0
msgpack==1.0.3
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
pathspec==0.11.2
pexpect==4.8.0
platformdirs==3.11.0
protobuf==4.21.12
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
sos==4.5.6
SQLAlchemy==2.0.23
ssh-import-id==5.11
starlette==0.27.0
systemd-python==234
texttable==1.6.4
tomli==2.0.1
Twisted==22.1.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.24.6
wadllib==1.3.6
watchfiles==0.21.0
websocket-client==1.2.3
websockets==12.0
zipp==1.0.0
zope.interface==5.4.0
